+++ go/pkg/pwcompose/const.go
@@ -1,0 +1,19 @@
+package pwcompose
+
+const (
+	labelPrefix           = "land.pathwar.compose."
+	serviceNameLabel      = labelPrefix + "service-name"
+	serviceOrigin         = labelPrefix + "origin"
+	challengeNameLabel    = labelPrefix + "challenge-name"
+	challengeVersionLabel = labelPrefix + "challenge-version"
+	InstanceKeyLabel      = labelPrefix + "instance-key"
+)
+
+const (
+	NginxContainerName = "pathwar-agent-nginx"
+	ProxyNetworkName   = "pathwar-proxy-network"
+)
+
+const (
+	defaultDockerPrefix = "pathwar/"
+)
